{{cookiecutter.project_slug}}/requirements.txt
Django>=5.0
django-environ==0.12.0
djangorestframework>=3.15
drf-spectacular>=0.27
psycopg2-binary>=2.9
python-decouple==3.8
PyJWT>=2.9
celery>=5.3
redis>=5.0
python-dotenv>=1.0